--- samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/cart/ui/AddressPickerState.kt
@@ -1,45 +1,0 @@
-/*
- * MIT License
- *
- * Copyright 2023-present, Shopify Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.ui
-
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import com.shopify.checkoutsheetkit.CheckoutAddressChangeRequestedEvent
-
-/**
- * Shared state holder for address change events.
- * Used to bridge between the event callback and Compose UI.
- */
-object AddressPickerState {
-    var currentEvent by mutableStateOf<CheckoutAddressChangeRequestedEvent?>(null)
-        private set
-
-    fun showPicker(event: CheckoutAddressChangeRequestedEvent) {
-        currentEvent = event
-    }
-
-    fun dismissPicker() {
-        currentEvent = null
-    }
-}
