--- samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/common/ui/CheckoutEventHandler.kt
@@ -1,61 +1,0 @@
-/*
- * MIT License
- *
- * Copyright 2023-present, Shopify Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package com.shopify.checkout_sdk_mobile_buy_integration_sample.common.ui
-
-import androidx.compose.runtime.Composable
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.ui.AddressPickerState
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.ui.AddressSelectionBottomSheet
-import com.shopify.checkoutsheetkit.CartDelivery
-import com.shopify.checkoutsheetkit.CartSelectableAddressInput
-import com.shopify.checkoutsheetkit.DeliveryAddressChangePayload
-
-/**
- * Handles checkout-related UI events (address picker, payment methods, etc.)
- * Centralizes all checkout event responses in one place.
- */
-@Composable
-fun CheckoutEventHandler() {
-    // Address picker bottom sheet
-    val addressEvent = AddressPickerState.currentEvent
-    addressEvent?.let { event ->
-        AddressSelectionBottomSheet(
-            onAddressSelected = { selectedAddress ->
-                event.respondWith(
-                    DeliveryAddressChangePayload(
-                        delivery = CartDelivery(
-                            addresses = listOf(
-                                CartSelectableAddressInput(
-                                    address = selectedAddress
-                                )
-                            )
-                        )
-                    )
-                )
-                AddressPickerState.dismissPicker()
-            },
-            onDismiss = {
-                AddressPickerState.dismissPicker()
-            }
-        )
-    }
-}
